Move & simplify rest api
Create new module for all REST related stuff.
Reorganize code:
1, move both routes to same file - 1 APIRouter() only
2, move pydantic models to another file
