+++ src/main.py
@@ -1,8 +1,7 @@
 from fastapi import FastAPI
 from strawberry.fastapi import GraphQLRouter
 
-from src import user_login
-from src import weather_conditions
+from src.rest_api.routes import router
 from src.dependencies.context import AppContext
 from src.schema import schema
 
@@ -11,5 +10,4 @@ app = FastAPI()
 graphql_app = GraphQLRouter(schema, context_getter=AppContext)
 
 app.include_router(graphql_app, prefix="/graphql")
-app.include_router(weather_conditions.router, prefix="")
-app.include_router(user_login.router, prefix="")
+app.include_router(router, prefix="")
